Rename slider-input to input-slider
