+++ projects/packages/forms/src/blocks/input-slider/icon.js
